Change nth-prime to use Swift Package Manager
