+++ exercises/nth-prime/Package.swift
@@ -1,0 +1,5 @@
+import PackageDescription
+
+let package = Package(
+    name: "nthPrime"
+)
