+++ exercises/nth-prime/Sources/nthPrime.swift
@@ -1,0 +1,1 @@
+
