+++ exercises/nth-prime/Tests/LinuxMain.swift
@@ -1,0 +1,6 @@
+import XCTest
+@testable import nthPrimeTests
+
+XCTMain([
+    testCase(nthPrimeTests.allTests),
+    ])
